+++ allennlp/training/callbacks/events.py
@@ -7,6 +7,10 @@ class Events:
     AFTER_BACKWARD = "AFTER_BACKWARD"
     BATCH_END = "BATCH_END"
 
+    BEFORE_VALIDATE = "BEFORE_VALIDATE"
+    VALIDATE = "VALIDATE"
+    AFTER_VALIDATE = "AFTER_VALIDATE"
+
     EPOCH_END = "EPOCH_END"
     TRAINING_END = "TRAINING_END"
